fix file names
